implement PR comments
